+++ NEWS.md
@@ -1,7 +1,8 @@
 # traktok 0.0.6
 
-* adds access to additional Research API endpoints ( tt_user_liked_videos_api, tt_user_pinned_videos_api, tt_user_follower_api, tt_user_following_api, tt_user_reposted_api, and tt_playlist_info_api)
+* adds access to additional Research API endpoints (tt_user_liked_videos_api, tt_user_pinned_videos_api, tt_user_follower_api, tt_user_following_api, tt_user_reposted_api, and tt_playlist_info_api)
 * tt_videos_hidden now supports Video IDs
+* adds tt_user_videos_api, a wrapper around tt_search_api to query user videos
 
 # traktok 0.0.5